用户状态确认，详情增加性别

+++ mtms/src/main/java/com/dili/mtms/api/AddressApi.java
@@ -7,7 +7,6 @@ import com.dili.mtms.service.AddressService;
 import com.dili.ss.constant.ResultCode;
 import com.dili.ss.domain.BaseOutput;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
@@ -33,11 +32,12 @@ public class AddressApi {
     @PostMapping(value="/listPage")
     public @ResponseBody BaseOutput listPage(Address address) {
         BaseData data = null;
+        //参数验证
+        if(address.getFirmId() == null || address.getCustomerId() == null){
+            return BaseOutput.create(ResultCode.DATA_ERROR,CfgContent.LACK_PARAM);
+        }
+
         try {
-            //参数验证
-            if(address.getFirmId() == null || address.getCustomerId() == null){
-                return BaseOutput.create(ResultCode.DATA_ERROR,"缺少必要参数");
-            }
             data = addressService.listAddress(address);
         }catch (Exception e){
             log.error(e.getMessage(),e);
@@ -53,11 +53,12 @@ public class AddressApi {
      */
     @PostMapping(value="/insert")
     public @ResponseBody BaseOutput insert(Address address) {
+        //参数验证
+        if(address.getFirmId() == null || address.getCustomerId() == null){
+            return BaseOutput.create(ResultCode.DATA_ERROR,CfgContent.LACK_PARAM);
+        }
+
         try {
-            //参数验证
-            if(address.getFirmId() == null || address.getCustomerId() == null){
-                return BaseOutput.create(ResultCode.DATA_ERROR,"缺少必要参数");
-            }
             int i = addressService.insertAddress(address);
         }catch (Exception e){
             log.error(e.getMessage(),e);
@@ -73,15 +74,17 @@ public class AddressApi {
      */
     @PostMapping(value="/update")
     public @ResponseBody BaseOutput update(Address address) {
+        //参数验证
+        if(address.getId() == null){
+            return BaseOutput.create(ResultCode.DATA_ERROR,CfgContent.LACK_PARAM);
+        }
+
         try {
+            //修改地址
             int i = addressService.updateSelective(address);
             //默认地址唯一
             if(i>0 && address.getIsDefault() != null && address.getCustomerId() != null){
                     addressService.updateIsdefaultAddress(address);
-            }
-            //地址修改失败
-            if(i<1){
-                return BaseOutput.create(ResultCode.DATA_ERROR,"修改失败");
             }
         }catch (Exception e){
             log.error(e.getMessage(),e);
@@ -97,11 +100,13 @@ public class AddressApi {
      */
     @PostMapping(value="/delete")
     public @ResponseBody BaseOutput delete(Address address) {
+        //参数验证
+        if(address.getId() == null){
+            return BaseOutput.create(ResultCode.DATA_ERROR,CfgContent.LACK_PARAM);
+        }
+
         try {
             int i = addressService.delete(address.getId());
-            if(i<1){
-                return BaseOutput.create(ResultCode.DATA_ERROR,"删除失败");
-            }
         }catch (Exception e){
             log.error(e.getMessage(),e);
             return BaseOutput.failure(CfgContent.SYSTEM_EXCEPTION);
@@ -117,6 +122,11 @@ public class AddressApi {
     @PostMapping(value="/getDefaultAddress")
     public @ResponseBody BaseOutput getDefaultAddress(Address address) {
         Address retrun_address = null;
+        //参数验证
+        if(address.getFirmId() == null || address.getCustomerId() == null){
+            return BaseOutput.create(ResultCode.DATA_ERROR,CfgContent.LACK_PARAM);
+        }
+
         try {
             retrun_address = addressService.getDefaultAddress(address);
         }catch (Exception e){
